advanced-api-project/requirements.txt
Django>=5.1.0
djangorestframework>=3.15.0
django-filter>=25.1
coreapi>=2.3.3
Markdown>=3.5.0